refactor: move add book button to OwnerFragment
