+++ app/src/main/java/com/cmput301f20t21/bookfriends/ui/add/AddEditActivity.java
@@ -23,7 +23,7 @@ import com.bumptech.glide.Glide;
 import com.cmput301f20t21.bookfriends.R;
 import com.cmput301f20t21.bookfriends.enums.BOOK_ACTION;
 import com.cmput301f20t21.bookfriends.enums.BOOK_ERROR;
-import com.cmput301f20t21.bookfriends.ui.library.LibraryFragment;
+import com.cmput301f20t21.bookfriends.ui.library.OwnedListFragment;
 import com.google.android.material.textfield.TextInputLayout;
 
 public class AddEditActivity extends AppCompatActivity {
@@ -85,7 +85,7 @@ public class AddEditActivity extends AppCompatActivity {
         });
 
         Intent intent = getIntent();
-        action = (BOOK_ACTION) intent.getSerializableExtra(LibraryFragment.BOOK_ACTION_KEY);
+        action = (BOOK_ACTION) intent.getSerializableExtra(OwnedListFragment.BOOK_ACTION_KEY);
     }
 
     @Override
