+++ app/src/main/java/com/cmput301f20t21/bookfriends/ui/library/OwnedListFragment.java
@@ -1,7 +1,7 @@
 package com.cmput301f20t21.bookfriends.ui.library;
 
+import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -14,10 +14,14 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.cmput301f20t21.bookfriends.R;
+import com.cmput301f20t21.bookfriends.enums.BOOK_ACTION;
+import com.cmput301f20t21.bookfriends.ui.add.AddEditActivity;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 public class OwnedListFragment extends Fragment {
-    private OwnedViewModel vm;
+    public static final String BOOK_ACTION_KEY = "com.cmput301f20t21.bookfriends.BOOK_ACTION";
 
+    private OwnedViewModel vm;
     private RecyclerView recyclerView;
     private RecyclerView.Adapter mAdapter;
     private RecyclerView.LayoutManager layoutManager;
@@ -27,12 +31,29 @@ public class OwnedListFragment extends Fragment {
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
                              @Nullable Bundle savedInstanceState) {
         vm = new ViewModelProvider(this).get(OwnedViewModel.class);
-        return inflater.inflate(R.layout.list_book, container, false);
+        View root = inflater.inflate(R.layout.owned_list_book, container, false);
+        final FloatingActionButton addBookButton = root.findViewById(R.id.add_button);
+
+        addBookButton.setOnClickListener(
+                view -> openAddEditActivity()
+        );
+
+        return root;
+    }
+
+    /**
+     * function allows user to jump into the add/edit screen when click on the floating button
+     */
+    private void openAddEditActivity() {
+        // TODO: Change the enum when calling the activity for editing
+        Intent intent = new Intent(this.getActivity(), AddEditActivity.class);
+        intent.putExtra(BOOK_ACTION_KEY, BOOK_ACTION.ADD);
+        startActivity(intent);
     }
 
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
-        recyclerView = (RecyclerView) view.findViewById(R.id.recycler_list_book);
+        recyclerView = (RecyclerView) view.findViewById(R.id.owned_recycler_list_book);
         // use this setting to improve performance if you know that changes
         // in content do not change the layout size of the RecyclerView
         recyclerView.setHasFixedSize(true);
