+++ app/src/main/res/layout/borrowed_list_book.xml
@@ -7,7 +7,7 @@
     >
 
     <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recycler_list_book"
+        android:id="@+id/borrowed_recycler_list_book"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:clipToPadding="false"
